moved code into components to avoid repetition
